// frontend/src/app/core/api/model/financial-contract-list-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FinancialContractListResponse { 
    id?: number;
    contractNumber?: string;
    promotionName?: string;
    studentFirstName?: string;
    studentLastName?: string;
    studentPhone?: string;
    guardianPhone?: string;
    totalAmount?: number;
    discountAmount?: number;
    netToPay?: number;
    totalCollected?: number;
    remainingDebt?: number;
    status?: FinancialContractListResponse.StatusEnum;
}
export namespace FinancialContractListResponse {
    export const StatusEnum = {
        Draft: 'DRAFT',
        Active: 'ACTIVE',
        FullyPaid: 'FULLY_PAID',
        Cancelled: 'CANCELLED',
        SettledWithDebt: 'SETTLED_WITH_DEBT'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}