// frontend/src/app/core/api/model/financial-contract-search-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FinancialContractSearchRequest { 
    keyword?: string;
    status?: FinancialContractSearchRequest.StatusEnum;
    promotionId?: number;
    trainingType?: FinancialContractSearchRequest.TrainingTypeEnum;
    page?: number;
    size?: number;
}
export namespace FinancialContractSearchRequest {
    export const StatusEnum = {
        Draft: 'DRAFT',
        Active: 'ACTIVE',
        FullyPaid: 'FULLY_PAID',
        Cancelled: 'CANCELLED',
        SettledWithDebt: 'SETTLED_WITH_DEBT'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const TrainingTypeEnum = {
        Continuous: 'CONTINUOUS',
        Accelerated: 'ACCELERATED',
        Accredited: 'ACCREDITED'
    } as const;
    export type TrainingTypeEnum = typeof TrainingTypeEnum[keyof typeof TrainingTypeEnum];
}