// frontend/src/app/core/api/model/activate-contract-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ActivateContractRequest { 
    discountAmount: number;
    paymentMethod: ActivateContractRequest.PaymentMethodEnum;
    paymentDate: string;
    reference?: string;
    referenceProvidedWhenRequired?: boolean;
}
export namespace ActivateContractRequest {
    export const PaymentMethodEnum = {
        Cash: 'CASH',
        Check: 'CHECK',
        BankTransfer: 'BANK_TRANSFER',
        CreditCard: 'CREDIT_CARD',
        OnlineStripe: 'ONLINE_STRIPE'
    } as const;
    export type PaymentMethodEnum = typeof PaymentMethodEnum[keyof typeof PaymentMethodEnum];
}